Add RPDB rules to bypass default interface rules
The bypassed rules are the rules for DEFAULT_NETWORK, DIRECTLY_CONNECTED and UNREACHABLE.
Similar rules are created, but with higher priority.

If the forced network goes down, the forced network will be unreachable.
If a network with the same interface as the previously forced network
goes up, the new network will be used as the forced network.